Add Log Module
The LogBootstrap is the starting point. It initialises a LogContext,
which binds LogSignal to LogCommand, and upon launch, dispatches a
LogSignal with “Hello World” message.

The LogCommand then gets triggered, and it logs the message in the
debug console.

This is what the LogModule does. Its a simple yet strange Hello World
example.