Add Persist Module
This one is a bit complicated from the Log module. The PersistBootstrap
initialises the PersistContext which then maps the IPersistenceService
to DiskPersistenceService, maps Delete, Load and Save signals to
Delete, Load and Save Commands respectively, and binds LoadedSignal to
itself since no command listens to it, just the PersistMediator.

PersistContext also binds PersistView to PersistMediator.

The Person value object stores the name and age. The IPerson interface
is present to not allow unauthorised entities in the module from
modifying the Person’s properties.

The PersistView dispatches local signals to its mediator, which then
dispatches appropriate signal to the module which then triggers
appropriate command and their functionality is executed.

The PersistMediator, as its name is, acts as the PersistView’s
mediator. The PersistView doesn’t know about the module. It just
dispatches input signals and tells what type of input has happened
using InputType enum.

The PersistMediator then handles those input signals. It also listens
to LoadedSignal which is dispatched by the DiskPersistenceService when
the Person object is successfully loaded from the disk. The
PersistMediator then updates the fields in PersistView appropriately.

Newtonsoft.Json is used to serialise and deserialise Person object and
PlayerPrefs is responsible for persisting the information.

The IPersistenceService is such that we can create
CloudPersistenceService and bind it with IPersistenceService, and the
data, instead of getting stored in disk will then be stored in cloud.