Add Echo Module
The Echo Module utilises WebSocket-Sharp to connect with an echo server
via web sockets and send messages to it.

The EchoBootstrap is the starting point, which initialises the
EchoContext. The EchoContext simply binds Connect, Disconnect and
SendMessage signals to Connect, Disconnect and SendMessage commands
respectively, EchoView to EchoMediator, IRoutineRunner to RoutineRunner
and IWebSocketService to EchoService.

The EchoView just dispatches when buttons are clicked, which then the
EchoMediator handles.

The EchoService is what it all comes down to. The WebSocket-Sharp runs
its code on a separate thread than the Unity’s main thread. Hence a
Producer-Consumer implementation is present in the EchoService.

Whenever the EchoSocket informs about some event, a string is added to
the LogQueue. This happens in a separate thread and not on the Unity
thread.

The LogQueue is continuously processed in the LogRoutine which runs as
a coroutine in the RoutineRunner object. This portion of code executes
in Unity’s main thread.

Since the LogQueue is a ConcurrentQueue, and a shared resource, one
thread produces logs which the other thread then consumes and
dispatches LogSignal.

The EchoMediator is listening for LogSignal and it acts accordingly,
displaying the results on screen.